refactored to remove header subfolder
The CI hates subfolders.
