+++ embedded/OH2_Lower_Instrument_Panel/2A13-BACKLIGHT_CONTROLLER/2A13-BACKLIGHT_CONTROLLER.h
@@ -27,7 +27,7 @@
  *   Non-Commercial - Share Alike License. (CC BY-NC-SA 4.0)
  *   ----------------------------------------------------------------------------------
  *   This Project uses Doxygen as a documentation generator.
- *   Please use Doxygen capable comments.
+ *   Please use Doxygen capable comments.d:\OpenHornet-Software\embedded\OH2_Lower_Instrument_Panel\2A13-BACKLIGHT_CONTROLLER\OH_BL_PANELS\1A2A1_MASTER_ARM_BL.h
  **************************************************************************************/
 /**
  * @file 2A13-BACKLIGHT_CONTROLLER.H
@@ -56,8 +56,8 @@
 
 Adafruit_NeoPixel uipBLCh1 = Adafruit_NeoPixel(UIP_LED_COUNT, UIP_BL_CH1, NEO_GRB + NEO_KHZ800);
 
-#include "OH_BL_PANELS\1A2A1_MASTER_ARM_BL.h"
-#include "OH_BL_PANELS\1A4_L_EWI_BL.h"
+#include "1A2A1_MASTER_ARM_BL.h"
+#include "1A4_L_EWI_BL.h"
 
 namespace OpenHornet {
 void setup() {
